+++ tests/__init__.py
@@ -2,7 +2,7 @@ def strict_eq(x, y):
     """Equality test bypassing the implicit string conversion in
     Python 2."""
     __tracebackhide__ = True
-    assert x == y
+    assert x == y, (x, y)
     assert issubclass(type(x), type(y)) or issubclass(type(y), type(x))
     if isinstance(x, dict) and isinstance(y, dict):
         x = sorted(x.items())
@@ -10,4 +10,4 @@ def strict_eq(x, y):
     elif isinstance(x, set) and isinstance(y, set):
         x = sorted(x)
         y = sorted(y)
-    assert repr(x) == repr(y)
+    assert repr(x) == repr(y), (x, y)